+++ src/Exception/DropFailedException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Symfony\Cmf\Component\Testing\Exception;
+
+/**
+ * @author Wouter J <wouter@wouterj.nl>
+ */
+class DropFailedException extends \RuntimeException
+{
+}